moved UseAnsiColoring into own module and added tests for get
